# docs/swagger.yaml
definitions:
  user.User:
    properties:
      age:
        type: integer
      birthday:
        type: string
      created_at:
        type: string
      firebase_uid:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /signup:
    post:
      consumes:
      - application/json
      description: get string by ID
      operationId: get-string-by-int2
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
        default:
          description: ""
          schema:
            type: integer
      summary: Signup a user
  /users/{id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      operationId: get-string-by-int
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
        default:
          description: ""
          schema:
            type: integer
      summary: Get a user
swagger: "2.0"